+++ tokio/tests/tracing_time.rs
@@ -17,6 +17,12 @@ async fn test_sleep_creates_span() {
         .named("runtime.resource")
         .with_target("tokio::time::sleep");
 
+    let poll_op = || {
+        expect::event()
+            .with_target("runtime::resource::poll_op")
+            .with_fields(expect::field("op_name").with_value(&"poll_elapsed"))
+    };
+
     let state_update = expect::event()
         .with_target("runtime::resource::state_update")
         .with_fields(
@@ -40,31 +46,35 @@ async fn test_sleep_creates_span() {
 
     let (subscriber, handle) = subscriber::mock()
         .new_span(sleep_span.clone().with_ancestry(expect::is_explicit_root()))
-        .enter(sleep_span.clone())
-        .event(state_update)
         .new_span(
             async_op_span
                 .clone()
-                .with_ancestry(expect::has_contextual_parent(&sleep_span_id))
+                .with_ancestry(expect::has_explicit_parent(&sleep_span_id))
                 .with_fields(expect::field("source").with_value(&"Sleep::new_timeout")),
         )
-        .exit(sleep_span.clone())
-        .enter(async_op_span.clone())
         .new_span(
             async_op_poll_span
                 .clone()
-                .with_ancestry(expect::has_contextual_parent(&async_op_span_id)),
+                .with_ancestry(expect::has_explicit_parent(&async_op_span_id)),
         )
+        .enter(sleep_span.clone())
+        .enter(async_op_span.clone())
+        .enter(async_op_poll_span.clone())
+        .event(poll_op())
+        .event(state_update)
+        .event(poll_op())
+        .exit(async_op_poll_span.clone())
+        .drop_span(async_op_poll_span)
         .exit(async_op_span.clone())
         .drop_span(async_op_span)
-        .drop_span(async_op_poll_span)
+        .exit(sleep_span.clone())
         .drop_span(sleep_span)
         .run_with_handle();
 
     {
         let _guard = tracing::subscriber::set_default(subscriber);
 
-        _ = tokio::time::sleep(Duration::from_millis(7));
+        tokio::time::sleep(Duration::from_millis(7)).await;
     }
 
     handle.assert_finished();